requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
CacheControl==0.14.1
cachelib==0.13.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
clickclick==20.10.2
cmake==3.31.1
connexion==2.14.2
cryptography==43.0.3
firebase-admin==6.6.0
Flask==2.2.5
Flask-Cors==5.0.0
Flask-Session==0.8.0
Flask-WTF==1.2.2
gcloud==0.18.3
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
lxml==5.3.0
MarkupSafe==3.0.2
msgpack==1.1.0
msgspec==0.18.6
oauth2client==4.1.3
outcome==1.3.0.post0
packaging==24.2
proto-plus==1.25.0
protobuf==5.29.0rc3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.0
pyparsing==3.2.0
Pyrebase4==4.8.0
PySocks==1.7.1
python-jwt==4.1.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==0.10.1
rpds-py==0.21.0
rsa==4.9
selenium==4.27.1
setuptools==58.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
swagger-ui-bundle==0.0.9
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
websocket-client==1.8.0
Werkzeug==2.2.3
wsproto==1.2.0
WTForms==3.2.1
yt-dlp==2024.11.18
pandas
openpyxl
seaborn
statsmodels